requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
h11==0.14.0
idna==3.4
outcome==1.2.0
Pillow==10.0.1
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
pytz==2023.3.post1
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==2.0.6
wsproto==1.2.0
